+++ packages/is/is.json/src/isJSON.spec.ts
@@ -1,17 +1,17 @@
 import { testData } from '../../../testValues';
-import { isJson } from './isJson';
+import { isJSON } from './isJSON';
 
 // extract the positive test cases, the rest will be negative
 const { arraysSafe, objectLiteralsSafe, ...rest } = testData;
 const positiveCases = [...arraysSafe, ...objectLiteralsSafe];
 const negativeCases = Object.values(rest).flat();
 
-describe('isJson', () => {
+describe('isJSON', () => {
   test.each(positiveCases)('should return TRUE for a/an %s', (title, value) => {
-    expect(isJson(value)).toEqual(true);
+    expect(isJSON(value)).toEqual(true);
   });
 
   test.each(negativeCases)('should return FALSE for a/an %s', (title, value) => {
-    expect(isJson(value)).toEqual(false);
+    expect(isJSON(value)).toEqual(false);
   });
 });