+++ packages/locale/locale.get/src/lib/formatLocale.ts
@@ -1,0 +1,5 @@
+export const formatLocale = (locale: string) => {
+  let [lang, region, ...rest] = locale.split('-');
+
+  return [lang.toLowerCase(), region.toUpperCase(), ...rest].filter((v) => v).join('-');
+};