+++ packages/locale/locale.get/src/lib/getIntlLocales.ts
@@ -1,0 +1,4 @@
+export const getIntlLocales = (): string[] => {
+  const locale = Intl.DateTimeFormat().resolvedOptions().locale;
+  return [locale];
+};