+++ packages/locale/locale/src/index.spec.ts
@@ -1,0 +1,7 @@
+import * as converters from './index';
+
+describe('Locale barrel package', () => {
+  it('should export specific packages and interfaces', () => {
+    expect(Object.keys(converters).sort()).toEqual(['localeGet', 'localeSettings']);
+  });
+});