+++ packages/locale/locale/src/index.ts
@@ -1,0 +1,2 @@
+export * from '@zerodep/locale.settings';
+export * from '@zerodep/locale.get';