+++ packages/to/to/src/index.ts
@@ -1,3 +1,4 @@
 export * from '@zerodep/to.errors';
 export * from '@zerodep/to.json';
+export * from '@zerodep/to.number';
 export * from '@zerodep/to.string';