+++ packages/types/types.currencies/src/index.ts
@@ -1,0 +1,2 @@
+export type { TypesCurrencies } from './TypesCurrencies';
+export type { TypesCurrenciesCrypto } from './TypesCurrenciesCrypto';