+++ packages/guard/guard/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Internal updates and refactorings
+
+## [0.4.0] - 2022-05-02
+
+### Changed
+
+- Fixed casing of the exported `guardJSON` utility
+
 ## [0.3.0] - 2022-05-02
 
 ### Added