+++ packages/types/types/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Introduce the `types.currencies` type declaration
+
 ## [0.1.0] - 2022-05-01
 
 ### Added
@@ -11,3 +17,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Introduce the `types` barrel package
 - Introduce the `types.locales` type declaration
 - Introduce the `types.timezones` type declaration
+- Introduce the `types.units` type declaration